gen_ai/project1_mcq_generator/requirements.txt
openai
langchain
langchain_community
streamlit
python-dotenv
PyPDF2
pandas



-e .